+++ ros-lunar/katana_moveit_ikfast_plugin/katana_moveit_ikfast_plugin-1.1.1.ebuild
@@ -7,14 +7,14 @@ DESCRIPTION="The katana_moveit_ikfast_plugin package"
 HOMEPAGE="http://wiki.ros.org/katana_moveit_ikfast_plugin"
 SRC_URI="https://github.com/uos-gbp/katana_driver-release/archive/release/lunar/katana_moveit_ikfast_plugin/1.1.1-0.tar.gz"
 
-LICENSE="UNKNOWN"
+LICENSE="||( BSD Apache )"
 KEYWORDS="~x86 ~amd64 ~arm ~arm64"
 
 RDEPEND="
-    ros-lunar/pluginlib
-    ros-lunar/roscpp
     ros-lunar/tf_conversions
+    ros-lunar/pluginlib
     ros-lunar/moveit_core
+    ros-lunar/roscpp
     virtual/lapack
 "
 DEPEND="${RDEPEND}
@@ -42,8 +42,10 @@ src_compile() {
 
 src_install() {
     cd ../../work
-    source /opt/ros/lunar/setup.bash
-    catkin_make_isolated --install --install-space="${D}/${ROS_PREFIX}" || die
-    rm -f ${D}/${ROS_PREFIX}/{.catkin,_setup_util.py,env.sh,setup.bash,setup.sh}
-    rm -f ${D}/${ROS_PREFIX}/{setup.zsh,.rosinstall}
+    source /${ROS_PREFIX}/setup.bash
+    /usr/bin/catkin_make_isolated --install --install-space="${D}/${ROS_PREFIX}" || die
+    if [[ -e ${D}/${ROS_PREFIX}/setup.bash ]]; then
+        rm -f ${D}/${ROS_PREFIX}/{.catkin,_setup_util.py,env.sh,setup.bash,setup.sh}
+        rm -f ${D}/${ROS_PREFIX}/{setup.zsh,.rosinstall}
+    fi
 }
