+++ ros-lunar/laser_pipeline/laser_pipeline-1.6.2.ebuild
@@ -3,8 +3,7 @@
 
 EAPI=6
 
-DESCRIPTION="Meta-package of libraries for processing laser data, including converting laser data
-      into 3D representations."
+DESCRIPTION="Meta-package of libraries for processing laser data, including converting laser "
 HOMEPAGE="http://www.ros.org/wiki/laser_pipeline"
 SRC_URI="https://github.com/ros-gbp/laser_pipeline-release/archive/release/lunar/laser_pipeline/1.6.2-0.tar.gz"
 
@@ -13,9 +12,9 @@ LICENSE="BSD"
 KEYWORDS="~x86 ~amd64 ~arm ~arm64"
 
 RDEPEND="
-    ros-lunar/laser_geometry
     ros-lunar/laser_filters
     ros-lunar/laser_assembler
+    ros-lunar/laser_geometry
 "
 DEPEND="${RDEPEND}
 "
@@ -42,8 +41,10 @@ src_compile() {
 
 src_install() {
     cd ../../work
-    source /opt/ros/lunar/setup.bash
-    catkin_make_isolated --install --install-space="${D}/${ROS_PREFIX}" || die
-    rm -f ${D}/${ROS_PREFIX}/{.catkin,_setup_util.py,env.sh,setup.bash,setup.sh}
-    rm -f ${D}/${ROS_PREFIX}/{setup.zsh,.rosinstall}
+    source /${ROS_PREFIX}/setup.bash
+    /usr/bin/catkin_make_isolated --install --install-space="${D}/${ROS_PREFIX}" || die
+    if [[ -e ${D}/${ROS_PREFIX}/setup.bash ]]; then
+        rm -f ${D}/${ROS_PREFIX}/{.catkin,_setup_util.py,env.sh,setup.bash,setup.sh}
+        rm -f ${D}/${ROS_PREFIX}/{setup.zsh,.rosinstall}
+    fi
 }
